+++ CHANGELOG.md
@@ -73,6 +73,7 @@
 ### Added
 - UI: added playlist revive tool to playlist contextual menu (within 'Other tools' submenu).
 - UI: added loaded status icon setting for UI-only playlists on non-lite mode; disabled by default. See [here](https://hydrogenaudio.org/index.php/topic,120979.msg1082334.html#msg1082334).
+- UI: added fallback to front art setting for Art by TF, if no file was found.
 - Exporting Playlist: added new Namida 'Export and convert' preset.
 - External integration: new callbacks.
 	* 'Playlist-Manager-SMP: switch enable panel'		-> { window: string[] }